backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
redis>=5.0.1
fastapi-cache2>=0.2.1
pydantic>=2.0.0
typing-extensions>=4.0.0
aioredis>=2.0.0